# docs/swagger.yaml
definitions:
  domain.BaseUser:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    required:
    - email
    - id
    - username
    type: object
...
  domain.FindTeamsResponse:
    properties:
      teams:
        items:
          $ref: '#/definitions/domain.Team'
        type: array
      total:
        type: integer
    required:
    - teams
    - total
    type: object
...
  domain.IDResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
    required:
    - id
    type: object
  domain.LoginResponse:
    properties:
      access_exp_time:
        example: "2026-03-05T09:50:42.108996Z"
        type: string
      access_token:
        example: sfjosdofj0293j4029jehsohsdofisdhoifhsdo
        type: string
      refresh_exp_time:
        example: "2026-03-05T09:50:42.108996Z"
        type: string
      refresh_token:
        example: sokdf11gpodjio23h9hsdoifhnso1321
        type: string
    required:
    - access_exp_time
    - access_token
    - refresh_exp_time
    - refresh_token
    type: object
  domain.Member:
    properties:
      id:
        type: string
      joined_at:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      user:
        $ref: '#/definitions/domain.BaseUser'
    required:
    - id
    - joined_at
    - role
    - user
    type: object
  domain.Role:
    enum:
    - admin
    - member
    - owner
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
    - RoleOwner
...
  domain.Team:
    properties:
      created_at:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/domain.Member'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/domain.BaseUser'
      updated_at:
        type: string
    required:
    - created_at
    - id
    - members
    - name
    - owner
    - updated_at
    type: object
  domain.TeamStats:
    properties:
      done_tasks_last_seven_days:
        type: integer
      id:
        type: string
      members_count:
        type: integer
      name:
        type: string
    required:
    - done_tasks_last_seven_days
    - id
    - members_count
    - name
    type: object
...
  domain.ValidatationErrors:
    properties:
      errors:
        items:
          $ref: '#/definitions/domain.ValidationError'
        type: array
    type: object
  domain.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  httpcommon.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  inputs.CreateTeam:
    properties:
      name:
        example: team 1
        type: string
    required:
    - name
    type: object
  inputs.CreateUser:
    properties:
      email:
        example: testemail@mail.ru
        type: string
      password:
        example: Assdfsdf2323!_
        type: string
      username:
        example: username123
        type: string
    required:
    - email
    - password
    - username
    type: object
  inputs.InviteMember:
    properties:
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
    required:
    - user_id
    type: object
  inputs.Login:
    properties:
      email:
        example: testemail@mail.ru
        type: string
      password:
        example: Assdfsdf2323!_
        type: string
    required:
    - email
    - password
    type: object
...
info:
  contact:
    email: f3d0t0tvmax@yandex.ru
    name: Fedotv Maxim (developer)
  description: Swagger Documentation for mkk_luna_test_rest_api (тестовое задание
    для компании "МКК ЛУНА")
  title: Swagger Documentation for mkk_luna_test_rest_api
  version: "1.0"
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Войти в аккаунт пользователя
      parameters:
      - description: Войти в аккаунт с помощью электронной почты и пароля
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.Login'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      summary: Войти в аккаунт
      tags:
      - users
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Создать новый аккаунт пользователя
      parameters:
      - description: Объект для создания аккаунта пользователя
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      summary: Создать новый аккаунт пользователя
      tags:
      - users
  /api/v1/statistics/teams:
    get:
      description: Возвращает статистику по всем командам
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TeamStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      summary: Статистика по командам
      tags:
      - statistics
...
  /api/v1/teams:
    get:
      consumes:
      - application/json
      description: олучить команды пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.FindTeamsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Получить команды пользователя
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Создать команду
      parameters:
      - description: Объект для создания команды
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateTeam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Создать команду
      tags:
      - teams
  /api/v1/teams/{id}/invite:
    post:
      consumes:
      - application/json
      description: Пригласить пользователя в команду
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Invite user to team dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.InviteMember'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Пригласить пользователя в команду
      tags:
      - teams
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer [ваш JWT токен]"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"